functions/geminiimgdesc/requirements.txt
functions-framework==3.*
google-cloud-aiplatform
google-cloud-datastore
google-cloud-translate
python-magic
Pillow
langchain
langsmith
langchain-google-vertexai